Refactor Kafka consumer code to use executor package for executing cron jobs
